+++ src/util.c
@@ -1,7 +1,7 @@
 #include "util.h"
-#include <string.h>
 
-int has_zip_extension(const char* filename)
+int8_t
+has_zip_extension(const char* filename)
 {
         const char* ext = strrchr(filename, '.');
         if (!ext || ext == filename)
